Create Shoe Store Database and Stats
Check out my shoe store database! First, I created a store database, displayed it to show shoes with a quantity of four or more, and ordered it by price. Lastly, I determined the total amount of shoes in the store and identified the highest and the lowest quantities in the store.