Superstore Analysis
Check out my Superstore Analysis! I first order all the items in the superstore by price. After that, I determined the average rating for merchandise in the store. Then, I determined the lowest price for Kitchen Supplies in the store. Lastly, I determined which items in the store had a stock quantity of 30 or less so that they could be re-ordered.